+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-19T03:59:27.771Z",
+    "2026-02-19T09:07:37.897Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {