+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-19T03:59:23.423Z",
+    "2026-02-19T09:07:33.562Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {